refactor: extract common function - hook - formatParamUrl - consumesFormOnly - plainJsonObjectToSwagger2 - localRefResolve
